requirements.txt
# Django
Django==5.0
python-dotenv==1.0.0
gunicorn==21.2.0

# WhatsApp
requests==2.31.0

# Google APIs
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==5.12.3
oauth2client==4.1.3

# AI/ML
google-generativeai==0.3.2

# PDF Processing
PyPDF2==3.0.1

# Utilities
Pillow==10.1.0
